Update and rename Project COOL.sh to ProjectCOOL.sh

+++ ProjectCOOL.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-# bash <(curl -s https://pages.dazerstudios.ga/Hoogeland-Menu/Project%20COOL.sh)
+# curl -s https://pages.dazerstudios.ga/Hoogeland-Menu/ProjectCOOL.sh | bash -s 1 1 1
 echo "        CCCCCCCCCCCCC     OOOOOOOOO          OOOOOOOOO     LLLLLLLLLLL             "
 echo "     CCC::::::::::::C   OO:::::::::OO      OO:::::::::OO   L:::::::::L             "
 echo "   CC:::::::::::::::C OO:::::::::::::OO  OO:::::::::::::OO L:::::::::L             "
@@ -27,16 +27,14 @@ echo "Installing Project COOL!"
 echo "Updating System"
 sudo apt update
 sudo apt upgrade -y
-clear
 echo "--------------------------------"
 echo "Done, updating!"
-echo "Installing Brave (Chrome but better)"
+echo "Installing Brave"
 sudo apt install apt-transport-https -y
 sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
 echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg arch=amd64] https://brave-browser-apt-release.s3.brave.com/ stable main"|sudo tee /etc/apt/sources.list.d/brave-browser-release.list
 sudo apt update
 sudo apt install brave-browser -y
-clear
 echo "--------------------------------"
 echo "Done installing Brave"
 echo "Changing a few settings for Brave"
@@ -45,25 +43,45 @@ mkdir ~/.config/BraveSoftware/Brave-Browser/
 echo "" > ~/.config/BraveSoftware/Brave-Browser/First\ Run
 mkdir ~/.config/BraveSoftware/Brave-Browser/Default/
 echo "$(curl -s https://pages.dazerstudios.ga/Hoogeland-Menu/Preferences)" > ~/.config/BraveSoftware/Brave-Browser/Default/Preferences
+if [ $1 = "1" ]
+then
 echo "Installing Discord"
 curl https://dl.discordapp.net/apps/linux/0.0.21/discord-0.0.21.deb --output discord.deb
 sudo apt install ./discord.deb -y
 rm discord.deb
-clear
 echo "--------------------------------"
 echo "Done installing Discord"
+fi
+if [ $2 = "1" ]
+then
 echo "Installing Spotify"
 curl -sS https://download.spotify.com/debian/pubkey_5E3C45D7B312C643.gpg | sudo apt-key add - 
 echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
 sudo apt update
 sudo apt install spotify-client -y
-clear
 echo "--------------------------------"
 echo "Done installing Spotify"
-echo "Opening Discord, Spotify and Brave"
+fi
+if [ $3 = "1" ]
+then
+echo "Installing Wine"
+sudo dpkg --add-architecture i386 
+sudo mkdir -pm755 /etc/apt/keyrings
+sudo curl https://dl.winehq.org/wine-builds/winehq.key --output /etc/apt/keyrings/winehq-archive.key
+sudo apt update
+sudo apt install --install-recommends winehq-stable
+echo "--------------------------------"
+echo "Done installing Wine"
+fi
 # Nohup runs the command in another process
 # So they can run next to eachother
 nohup brave-browser >/dev/null 2>&1 &
+if [ $1 = "1" ]
+then
 nohup discord >/dev/null 2>&1 &
+fi
+if [ $2 = "1" ]
+then
 nohup spotify >/dev/null 2>&1 &
+fi
 echo "Done installing Project COOL!"
